luci-app-attendedsysupgrade: reorder and reword error instructions
Many users have been posting incomplete error information to the
forum and in git issues.  They'll not report important information
that is above the paragraph indicating where to report it, only
the partial information below it.

Remedy this by putting the reporting instructions first, and make
explicit that everything below is needed.  Add a new button that
creates a report appropriate for a forum post and saves it to the
clipboard.

Signed-off-by: Eric Fahlgren <ericfahlgren@gmail.com>
(cherry picked from commit e2f02198f9010152022ad15ffe2635fc490bf96a)