luci-mod-network: show stop button if interface start is pending
Currently, it is not possible to stop an interface from establishing a
connection when it is flageed as pending in the netifd. This situation
occurs when netifd tells a proto to establish a connection, but the
proto is executing a blocken binary. This situation occurs in the proto
'modemmanager' when it wants to register but cannot. The timeout is set to
'120' seconds. During this time, the establishment cannot be stopped via
LuCI because the disable button can not be clicked.

To fix this, the pending flag is also evaluated and the disable button
is displayed in this state as well.

Signed-off-by: Florian Eckert <fe@dev.tdt.de>