luci-mod-network: organize "limits" tab in dhcp/dns views
Sort the options in the "limits" tab so that they are all grouped together in
the order they appear in the UI and delete ones that are specific to DNS from
dhcp.js and vice versa.

Only one option is relevant to the dhcp view, so move it to the "general" tab.

Signed-off-by: David Härdeman <david@hardeman.nu>