luci-app-attendedsysupgrade: use the same logic as the main build when requesting rebuilds
When initiating a build the requests were build differently
between main and rebuild servers leading to the latter using
only POST request instead of the POST->GET flow. This led to
incorrect tracking of builds in the target ASU server due to
each request being classified unique.

This is fixed by switching rebuild requests to the POST->GET
flow and storing each server's request hash in a map and using
it regardless if it is the main or a rebuild server handling
the request.

Fixes: openwrt/asu#1526
Signed-off-by: Lili González <lili@guerra24.net>
(cherry picked from commit 31fb9ce1967ee5650777588efaa167a6ac076417)