luci-app-attendedsysupgrade: improve ux during install and reboot
During the sysupgrade image installation and reboot process, it
is possible for LuCI to reconnect before the device is rebooted.
This leads to confusion and possible danger of changes being made
that will damage the device.

We set expectations as to what will happen during the installation
process with better wording in the spin dialog, and also "lock up"
the interface for 10s, allowing sysupgrade to do its job and avoid
reconnecting too early

Link: https://forum.openwrt.org/t/luci-attended-sysupgrade-support-thread/230552/268
Signed-off-by: Eric Fahlgren <ericfahlgren@gmail.com>