luci-app-attendedsysupgrade: do major cleanup of error handler
Extract and isolate the forum-specific constants and formats to
a location that is easily modified.  Factor out a lot of common
code and simply make everything much cleaner and more readable.

Output to the "Request Data" block was augmented to contain the
system board info, and was reordered to put more important data
earlier in the output.

Also fix a potential bug in construction of the error html, where
the final step was coercing everything to a string prematurely.

Signed-off-by: Eric Fahlgren <ericfahlgren@gmail.com>
(cherry picked from commit cf868f439855534f62f2cfbd2732811d53a26c25)