rpcd-mod-luci: correct DUID logic
dnsmasq writes a statefile for DHCPv4 leases with the MAC address as the
second field and the clientid as the fifth field. The clientid starts
with a type byte (0x01 = ethernet, 0xff = DUID/IAID identifier).

Similarly, odhcpd writes the MAC address as the second field for IPv4
leases.

This patch changes the behaviour of rpcd-mod-luci so that it doesn't
report a non-DUID as a DUID.

Signed-off-by: David Härdeman <david@hardeman.nu>