luci-app-watchcat: revert pinghosts changes
This change introduced support for multiple ping hosts, but due to
inconsistent argument handling with the backend, it results in broken
behavior.

As part of this revert, the pinghosts option is disabled in LUCI to
restore the previous, known-working behavior. This aligns LUCI with the
current backend until a proper fix is implemented.

Signed-off-by: Ivan Diaz <diaz.it@icloud.com>